+++ apollo-router/src/plugins/telemetry/config_new/supergraph/events.rs
@@ -1,0 +1,1 @@
+
