+++ apollo-router/src/plugins/telemetry/config_new/supergraph/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod attributes;
+pub(crate) mod events;
+pub(crate) mod instruments;
+pub(crate) mod selectors;
+pub(crate) mod spans;
